Extract DownloadAndValidateFile to FileCache.
